Added RelativeContainer unit test. Split out RelativeContainer into core.
