+++ source/core/core.vcxproj
@@ -191,6 +191,7 @@
     <ClInclude Include="slang-platform.h" />
     <ClInclude Include="slang-process-util.h" />
     <ClInclude Include="slang-random-generator.h" />
+    <ClInclude Include="slang-relative-container.h" />
     <ClInclude Include="slang-render-api-util.h" />
     <ClInclude Include="slang-secure-crt.h" />
     <ClInclude Include="slang-shared-library.h" />
@@ -220,6 +221,7 @@
     <ClCompile Include="slang-object-scope-manager.cpp" />
     <ClCompile Include="slang-platform.cpp" />
     <ClCompile Include="slang-random-generator.cpp" />
+    <ClCompile Include="slang-relative-container.cpp" />
     <ClCompile Include="slang-render-api-util.cpp" />
     <ClCompile Include="slang-shared-library.cpp" />
     <ClCompile Include="slang-std-writers.cpp" />
