+++ source/core/slang-relative-container.h
@@ -1,0 +1,175 @@
+// slang-relative-container.h
+#ifndef SLANG_RELATIVE_CONTAINER_H_INCLUDED
+#define SLANG_RELATIVE_CONTAINER_H_INCLUDED
+
+#include "slang-basic.h"
+
+namespace Slang {
+
+enum
+{
+    kNullOffset32 = 0x80000000
+};
+
+class RelativeContainer;
+
+template <typename T>
+class Safe32Ptr
+{
+public:
+    typedef Safe32Ptr ThisType;
+
+    T& operator*() const { return *get(); }
+    T* operator->() const { return get(); }
+    operator T*() const { return get(); }
+
+    Safe32Ptr(const ThisType& rhs) : m_offset(rhs.m_offset), m_container(rhs.m_container) {}
+    const Safe32Ptr& operator=(const ThisType& rhs) { m_offset = rhs.m_offset; return *this; }
+
+    SLANG_FORCE_INLINE T* get() const;
+
+    Safe32Ptr() : m_container(nullptr), m_offset(kNullOffset32) {}
+
+    Safe32Ptr(int32_t offset, RelativeContainer* container) : m_offset(offset), m_container(container) {}
+
+    RelativeContainer* m_container;
+    int32_t m_offset;
+};
+
+template <typename T>
+class Relative32Ptr
+{
+public:
+    typedef Relative32Ptr ThisType;
+
+    T& operator*() const { return *get(); }
+    T* operator->() const { return get(); }
+    operator T*() const { return get(); }
+
+    T* get()
+    {
+        uint8_t* nonConstThis = (uint8_t*)this;
+        return (m_offset == kNullOffset) ? nullptr : (T*)(nonConstThis + m_offset);
+    }
+    T* get() const
+    {
+        uint8_t* nonConstThis = const_cast<uint8_t*>((const uint8_t*)this);
+        return (m_offset == kNullOffset32) ? nullptr : (T*)(nonConstThis + m_offset);
+    }
+
+    T* detach() { T* ptr = get(); m_offset = kNullOffset32; }
+
+    void setNull() { m_offset = kNullOffset32; }
+
+    SLANG_FORCE_INLINE void set(T* ptr) { m_offset = ptr ? int32_t(((uint8_t*)ptr) - ((const uint8_t*)this)) : uint32_t(kNullOffset32); }
+
+    Relative32Ptr(const Safe32Ptr<T>& rhs) { set(rhs.get()); }
+    Relative32Ptr() :m_offset(kNullOffset) {}
+    Relative32Ptr(T* ptr) { set(ptr); }
+
+    const Relative32Ptr& operator=(const ThisType& rhs) { m_offset = rhs.m_offset; }
+    const Relative32Ptr& operator=(const Safe32Ptr& rhs) { set(rhs.get()); }
+
+    int32_t m_offset;
+};
+
+template <typename T>
+class Relative32Array
+{
+    Relative32Array() :m_count(0) {}
+    uint32_t m_count;               ///< the size of the data
+    Relative32Ptr m_data;           ///< The data
+};
+
+struct RelativeString
+{
+    enum
+    {
+        kSizeBase = 252,
+        kMaxSizeEncodeSize = 5,
+    };
+
+    /// Get contents as a slice
+    UnownedStringSlice getSlice() const;
+    /// Get null terminated string
+    const char* getCstr() const;
+
+    /// Decode the size. Returns the start of the string text, and outSize holds the size (NOT including terminating 0)
+    static const char* decodeSize(const char* in, size_t& outSize);
+
+    /// Returns the amount of bytes used, end encoding in 'encode'
+    static size_t calcEncodedSize(size_t size, uint8_t encode[kMaxSizeEncodeSize]);
+    /// Calculate the total size needed to store the string *including* terminating 0
+    static size_t calcAllocationSize(const UnownedStringSlice& slice);
+
+    /// Calculate the total size needed to store string. Size should be passed *without* terminating 0
+    static size_t calcAllocationSize(size_t size);
+
+    char m_sizeThenContents[1];
+};
+
+class RelativeContainer
+{
+public:
+
+    template <typename T>
+    Safe32Ptr<T> allocate()
+    {
+        void* data = allocate(sizeof(T), SLANG_ALIGN_OF(T));
+        memset(data, 0, size);
+        return Safe32Ptr<T>(getOffset(data), this);
+    }
+
+    /// Allocate without alignment (effectively 1)
+    void* allocate(size_t size);
+    void* allocate(size_t size, size_t alignment);
+    void* allocateAndZero(size_t size, size_t alignment);
+
+    template <typename T>
+    T* toPtr(Safe32Ptr<T> ptr)
+    {
+        if (ptr.m_offset == kNullOffset)
+        {
+            return nullptr;
+        }
+        SLANG_ASSERT(ptr.m_offset >= 0 && ptr.m_offset <= m_current)
+            return (T*)(m_data.getBuffer() + ptr.m_offset);
+    }
+
+    template <typename T>
+    Safe32Ptr<T> toSafe(T* ptr) { SafePtr<T> safePtr; relPtr.m_offset = getOffset(); }
+    int32_t getOffset(const void* ptr)
+    {
+        ptrdiff_t offset = ((const uint8_t*)ptr) - m_data.getBuffer();
+        if (offset < 0 || size_t(offset) > m_current)
+        {
+            return int32_t(kNullOffset32);
+        }
+        return int32_t(offset);
+    }
+
+    Safe32Ptr<RelativeString> newString(const UnownedStringSlice& slice);
+    Safe32Ptr<RelativeString> newString(const char* contents);
+
+    /// Get the contained data
+    uint8_t* getData() { return m_data.getBuffer(); }
+
+    /// Ctor
+    RelativeContainer();
+
+protected:
+    size_t m_current;
+    List<uint8_t> m_data;
+};
+
+// --------------------------------------------------------------------------
+
+template <typename T>
+SLANG_FORCE_INLINE T* Safe32Ptr<T>::get() const
+{
+    return m_container ? ((T*)(m_container->getData() + m_offset)) : (T*)nullptr;
+}
+
+} // namespace Slang
+
+#endif
